Attempt to make BattleManager Functional
Added dummy functionality to UI Elements. Added Debug lines to see where script terminates. Script was being caught in quite a few infinte for and if statements. Commented out GameObject. Find on UI elements as per the Unity API this can only be done as the GameObject "must not have a parent in the Hierarchy view"